+++ src/all/test.ts
@@ -4,7 +4,7 @@ import set from '../set'
 import { collection } from '../collection'
 import { Ref, ref } from '../ref'
 import get from '../get'
-import clear from '../clear'
+import remove from '../remove'
 
 describe('all', () => {
   type Book = { title: string }
@@ -18,8 +18,8 @@ describe('all', () => {
       set(books, 'sapiens', { title: 'Sapiens' }),
       set(books, '22laws', { title: 'The 22 Immutable Laws of Marketing' }),
       set(books, 'momtest', { title: 'The Mom Test' }),
-      clear(books, 'hp1'),
-      clear(books, 'hp2')
+      remove(books, 'hp1'),
+      remove(books, 'hp2')
     ])
   })